+++ src/lib/ericchase/Platform/Node/Path.ts
@@ -1,25 +1,29 @@
-import { default as node_path } from 'node:path';
+import node_path from 'node:path';
 import { PrepareMessage } from '../../Utility/PrepareMessage.js';
 
-export function JoinPaths(...paths: string[]) {
-  return node_path.join(...paths);
+/** Use `new Path(...).appendSegment(...)` instead. */
+export function JoinPaths(...paths: (Path | PathGroup)[]) {
+  return node_path.join(...paths.map((path) => path.path));
 }
 
-export function NormalizePath(path: string) {
-  return node_path.normalize(path);
+/** Use `new Path(...).path` instead. */
+export function NormalizePath(path: Path | PathGroup) {
+  return node_path.normalize(path.path);
 }
 
-export function ParsePath(path: string) {
-  return node_path.parse(path);
+/** Use `new Path(...)` instead. */
+export function ParsePath(path: Path | PathGroup) {
+  return node_path.parse(path.path);
 }
 
-export function ResolvePath(path: string) {
-  return node_path.resolve(path);
+/** Use `new Path(...).resolve` instead. */
+export function ResolvePath(path: Path | PathGroup) {
+  return node_path.resolve(path.path);
 }
 
-// ! not tested
-export function SanitizePath(path: string) {
-  return path.replace(/[^a-z0-9\.\_\-]/gi, '_').toLowerCase();
+/** Use `new Path(...).sanitize` instead. */
+export function SanitizePath(path: Path | PathGroup) {
+  return path.path.replace(/[^a-z0-9\.\_\-]/gi, '_').toLowerCase();
 }
 
 export const PathSeparator = node_path.sep;
@@ -43,27 +47,31 @@ export class Path {
     this.$standard_path = this.path.split(node_path.sep).join('/');
   }
   static build({ dir = '', base = '' }) {
-    if (dir.length === 0) {
-      return new Path(base);
-    }
-    if (base.length === 0) {
-      return new Path(dir);
-    }
-    return new Path(`${dir}/${base}`);
+    return new Path(dir).appendSegment(base);
   }
-  static from(pathOrString: Path | string) {
+  static from(pathOrString: Path | PathGroup | string) {
     if (typeof pathOrString === 'string') {
       return new Path(pathOrString);
+    }
+    if (pathOrString instanceof PathGroup) {
+      return new Path(pathOrString.path);
     }
     return pathOrString;
   }
   get path() {
     return this.$path;
   }
+  get resolve() {
+    return node_path.resolve(this.path);
+  }
+  /** ! Not Tested ! */
+  get sanitize() {
+    return this.path.replace(/[^a-z0-9\.\_\-]/gi, '_').toLowerCase();
+  }
   get standard_path() {
     return this.$standard_path;
   }
-  appendSegment(pathOrString: Path | string) {
+  appendSegment(pathOrString: Path | PathGroup | string) {
     return new Path(node_path.join(this.path, Path.from(pathOrString).path));
   }
   newDir(new_dir: string) {
@@ -107,6 +115,9 @@ export class Path {
     }
     return this.newBase(`${this.name}.${new_ext}`);
   }
+  toString() {
+    return this.path;
+  }
 }
 
 export class PathSet {
@@ -114,11 +125,13 @@ export class PathSet {
   get paths() {
     return this.pathIterator();
   }
-  add(path: Path) {
+  add(pathOrString: Path | PathGroup | string) {
+    const path = Path.from(pathOrString);
     this.path_map.set(path.path, path);
     return this;
   }
-  has(path: Path) {
+  has(pathOrString: Path | PathGroup | string) {
+    const path = Path.from(pathOrString);
     return this.path_map.has(path.path);
   }
   *pathIterator() {
@@ -129,30 +142,31 @@ export class PathSet {
 }
 
 export class PathGroup {
-  readonly origin: string; // includes root, but not dir
+  readonly origin_path: Path;
+  readonly origin_dir: string; // includes root, but not dir
   readonly root: string;
-  readonly dir: string;
-  readonly base: string; // includes name and ext
-  readonly name: string;
-  readonly ext: string;
+  readonly relative_path: Path;
+  readonly relative_dir: string; // does not include root
+  readonly relative_base: string; // includes name and ext
+  readonly relative_name: string;
+  readonly relative_ext: string;
   readonly $path: string;
   readonly $standard_path: string;
-  constructor(
-    readonly origin_path: Path,
-    readonly relative_path: Path,
-  ) {
-    this.relative_path = relative_path.newRoot('');
-    this.origin = origin_path.path;
-    this.root = origin_path.root;
-    this.dir = relative_path.dir;
-    this.base = relative_path.base;
-    this.name = relative_path.name;
-    this.ext = relative_path.ext;
-    this.$path = node_path.join(this.origin_path.path, this.relative_path.path);
-    this.$standard_path = this.$path.split(node_path.sep).join('/');
+  constructor(origin_path: Path | PathGroup | string, relative_path: Path | PathGroup | string) {
+    this.origin_path = Path.from(origin_path);
+    this.relative_path = Path.from(relative_path).newRoot('');
+    this.origin_dir = this.origin_path.path;
+    const final_path = Path.from(this.origin_path).appendSegment(this.relative_path);
+    this.$path = final_path.path;
+    this.$standard_path = final_path.standard_path;
+    this.root = final_path.root;
+    this.relative_dir = final_path.dir;
+    this.relative_base = final_path.base;
+    this.relative_name = final_path.name;
+    this.relative_ext = final_path.ext;
   }
-  static build({ origin_path = '', relative_path = '' }) {
-    return new PathGroup(new Path(origin_path), new Path(relative_path));
+  static build({ origin_path = '', relative_path = '' }: { origin_path?: Path | PathGroup | string; relative_path?: Path | PathGroup | string }) {
+    return new PathGroup(origin_path, relative_path);
   }
   get path() {
     return this.$path;
@@ -160,20 +174,26 @@ export class PathGroup {
   get standard_path() {
     return this.$standard_path;
   }
-  newOrigin(new_origin_path: Path | string) {
-    return new PathGroup(Path.from(new_origin_path), this.relative_path);
+  newOrigin(new_origin_path: Path | PathGroup | string) {
+    return new PathGroup(new_origin_path, this.relative_path);
   }
-  newDir(new_dir: string) {
+  newRelative(new_relative_path: Path | PathGroup | string) {
+    return new PathGroup(this.origin_path, new_relative_path);
+  }
+  newRelativeDir(new_dir: string) {
     return new PathGroup(this.origin_path, this.relative_path.newDir(new_dir));
   }
-  newBase(new_base: string) {
+  newRelativeBase(new_base: string) {
     return new PathGroup(this.origin_path, this.relative_path.newBase(new_base));
   }
-  newName(new_name: string) {
+  newRelativeName(new_name: string) {
     return new PathGroup(this.origin_path, this.relative_path.newName(new_name));
   }
-  newExt(new_ext: string) {
+  newRelativeExt(new_ext: string) {
     return new PathGroup(this.origin_path, this.relative_path.newExt(new_ext));
+  }
+  toString() {
+    return this.path;
   }
 }