+++ tools/lib/format.ts
@@ -1,7 +1,7 @@
 import { RunSync } from '../../src/lib/ericchase/Platform/Bun/Child Process.js';
 import { command_map } from '../dev.js';
-import { Cache_FileStats_Lock, Cache_FileStats_Unlock } from '../lib/cache/FileStatsCache.js';
-import { TryLock } from '../lib/cache/LockCache.js';
+import { Cache_FileStats_Lock, Cache_FileStats_Unlock } from './cache/FileStatsCache.js';
+import { TryLock } from './cache/LockCache.js';
 
 TryLock(command_map.format);