+++ tools/lib/hot_reload.ts
@@ -1,4 +1,3 @@
-import { server_http } from '../../src/dev_server/server-data.js';
 import { ConsoleError, ConsoleLog } from '../../src/lib/ericchase/Utility/Console.js';
 import { Debouncer } from '../../src/lib/ericchase/Utility/Debounce.js';
 import { on_log } from '../scripts/build.js';
@@ -7,9 +6,11 @@ export class HotReloader {
   enabled = false;
   hotreload: Debouncer<Promise<void>>;
   unsubscribe = () => {};
-  constructor(debounce_delay = 250) {
+  readonly $reload_endpoint: string;
+  constructor(reload_endpoint: string, debounce_delay = 250) {
+    this.$reload_endpoint = reload_endpoint;
     this.hotreload = new Debouncer(async () => {
-      await fetch(`${server_http}/server/reload`);
+      await fetch(reload_endpoint);
     }, debounce_delay);
   }
   disable() {
@@ -23,7 +24,8 @@ export class HotReloader {
   async enable() {
     if (this.enabled === false) {
       try {
-        await fetch(server_http);
+        const url = new URL(this.$reload_endpoint);
+        await fetch(url.host);
         this.enabled = true;
         this.unsubscribe = on_log.subscribe(() => this.hotreload.run());
         ConsoleError('Hot Reloading: ONLINE');