+++ tools/lib/preprocessors/FilePreprocessor.ts
@@ -1,6 +1,7 @@
+import type { Path, PathGroup } from '../../../src/lib/ericchase/Platform/Node/Path.js';
 import type { SyncAsync } from '../../../src/lib/ericchase/Utility/Types.js';
 
 export interface FilePreprocessor {
-  pathMatches(path: string): boolean;
+  pathMatches(path: Path | PathGroup): boolean;
   preprocess(content: string): SyncAsync<{ content: string }>;
 }