+++ tools/scripts/build.ts
@@ -3,12 +3,11 @@ import { RunSync } from '../../src/lib/ericchase/Platform/Bun/Child Process.js';
 import { GlobScanner } from '../../src/lib/ericchase/Platform/Bun/Glob.js';
 import { CleanDirectory, RenameFile } from '../../src/lib/ericchase/Platform/Node/Fs.js';
 import { Path } from '../../src/lib/ericchase/Platform/Node/Path.js';
-import { ConsoleLog, ConsoleNewline } from '../../src/lib/ericchase/Utility/Console.js';
+import { ConsoleLogWithDate, ConsoleNewline } from '../../src/lib/ericchase/Utility/Console.js';
 import { command_map } from '../dev.js';
 import { BuildRunner, copy, IntoPatterns, processHTML } from '../lib/build.js';
 import { Cache_FileStats_Lock, Cache_FileStats_Reset, Cache_FileStats_Unlock } from '../lib/cache/FileStatsCache.js';
 import { Cache_Unlock, TryLock, TryLockEach } from '../lib/cache/LockCache.js';
-import { IIFEWrapperPreprocessor } from '../lib/preprocessors/FilePreprocessor_IIFEWrapper.js';
 import { CustomComponentPreprocessor } from '../lib/preprocessors/HTMLPreprocessor_CustomComponent.js';
 import { ImportConverterPreprocessor } from '../lib/preprocessors/HTMLPreprocessor_ImportConverter.js';
 
@@ -35,15 +34,12 @@ const component_patterns = ['components/**/*.html'];
 const dot_component_patterns = ['components/**/.*.html']; // not included in regular scans
 
 // preprocessors
-const iife_wrapper = new IIFEWrapperPreprocessor(script_suffixes.map((suffix) => `${suffix}.js`));
+// const iife_wrapper = new IIFEWrapperPreprocessor(script_suffixes.map((suffix) => `${suffix}.js`));
 const custom_component = new CustomComponentPreprocessor();
 const import_converter = new ImportConverterPreprocessor(
-  ...IntoPatterns([...module_suffixes, ...script_suffixes], source_extensions).map((pattern) => [pattern, '.js'] as const), //
-  ...source_extensions.map((ext) => [ext, '.js'] as const),
+  ...source_extensions.map((ext) => [ext, '.js'] as const), //
 );
-const file_renamer: (readonly [string, string])[] = [
-  ...IntoPatterns([...module_suffixes, ...script_suffixes], '.js').map((pattern) => [pattern, '.js'] as const), //
-];
+const file_renamer: (readonly [string, string])[] = [];
 
 // build mode
 export const build_mode = {
@@ -52,41 +48,36 @@ export const build_mode = {
 };
 
 // bundler
-const bundler = new BuildRunner(onLog);
+const bundler = new BuildRunner();
+bundler.broadcast.subscribe(() => {
+  for (const line of bundler.output) {
+    if (line.length > 0) {
+      onLog(`bund: ${line}`);
+    }
+  }
+  bundler.output = [];
+});
 
 // step: clean
 export async function buildStep_Clean() {
-  ConsoleLog('Build Step: Clean');
   bundler.killAll();
   Cache_FileStats_Reset();
-  await CleanDirectory(out_dir.path);
+  await CleanDirectory(out_dir);
   // await CleanDirectory(tmp_dir.path);
 }
 
-// step: bundle
+// step: setup bundler
 export async function buildStep_SetupBundler() {
-  ConsoleLog('Build Step: Setup Bundler');
+  const sourcemap = build_mode.watch === true ? 'inline' : 'linked';
+  const external = [...external_import_patterns];
+  const watch = build_mode.watch;
   // modules
   for (const entry of new GlobScanner().scan(src_dir, ...module_patterns).path_groups) {
-    bundler.add({
-      entry,
-      external_imports: [...external_import_patterns],
-      out_dir: out_dir,
-      sourcemap_mode: 'linked',
-      target: 'browser',
-      watch: build_mode.watch,
-    });
+    bundler.add({ entrypoint: entry, outdir: out_dir, sourcemap, external, watch });
   }
   // scripts
   for (const entry of new GlobScanner().scan(src_dir, ...script_patterns).path_groups) {
-    bundler.add({
-      entry,
-      external_imports: [],
-      out_dir: out_dir,
-      sourcemap_mode: 'linked',
-      target: 'browser',
-      watch: build_mode.watch,
-    });
+    bundler.add({ entrypoint: entry, outdir: out_dir, sourcemap, external: [], watch });
   }
   if (build_mode.watch === false) {
     await Promise.allSettled([...bundler.subprocess_map].map(([_, process]) => process.exited));
@@ -96,12 +87,12 @@ export async function buildStep_SetupBundler() {
 // step: process html
 export async function buildStep_ProcessHTMLFiles() {
   // add components
-  for (const { name, path } of new GlobScanner().scan(src_dir, ...component_patterns).path_groups) {
-    custom_component.registerComponentPath(name, path);
+  for (const path_group of new GlobScanner().scan(src_dir, ...component_patterns).path_groups) {
+    custom_component.registerComponentPath(Path.from(path_group).name, path_group);
   }
   // add dot components (notice `scanDot` and `registerComponentPath(,,true)`
-  for (const { name, path } of new GlobScanner().scanDot(src_dir, ...dot_component_patterns).path_groups) {
-    custom_component.registerComponentPath(name.slice(1), path, true);
+  for (const path_group of new GlobScanner().scanDot(src_dir, ...dot_component_patterns).path_groups) {
+    custom_component.registerComponentPath(Path.from(path_group).name.slice(1), path_group, true);
   }
 
   const processed_html_paths = await processHTML({
@@ -111,7 +102,7 @@ export async function buildStep_ProcessHTMLFiles() {
     preprocessors: [custom_component, import_converter],
   });
   for (const path of processed_html_paths.paths) {
-    onLog(`Preprocessed: ${path}`);
+    onLog(`html: ${path}`);
   }
 
   if (build_mode.watch === false) {
@@ -124,23 +115,22 @@ export async function buildStep_ProcessHTMLFiles() {
 
 // step: copy
 export async function buildStep_Copy() {
-  ConsoleLog('Build Step: Copy');
-  const copied_pathset = await copy({
+  const src_copied_paths = await copy({
     out_dirs: [out_dir],
     to_copy: new GlobScanner().scan(src_dir, '**/*'),
     to_exclude: new GlobScanner().scan(src_dir, '**/*.html', ...component_patterns, ...lib_patterns, ...module_patterns, ...script_patterns, ...source_patterns),
   });
-  // const tmp_copied_pathset = await copy({
+  // const tmp_copied_paths = await copy({
   //   out_dirs: [out_dir],
   //   to_copy: new GlobScanner().scan(tmp_dir, '**/*'), // exclude nothing
   //   preprocessors: [iife_wrapper],
   // });
   const copied_paths = new Set([
-    ...copied_pathset.paths, //
-    // ...tmp_copied_pathset.paths,
+    ...src_copied_paths.paths, //
+    // ...tmp_copied_paths.paths,
   ]);
   for (const path of copied_paths) {
-    onLog(`Copied: ${path}`);
+    onLog(`copy: ${path}`);
   }
   if (build_mode.watch === false) {
     onLog(`${copied_paths.size} files copied.`);
@@ -178,11 +168,11 @@ export async function buildStep_Move() {
 
 // step: rename
 export async function buildStep_Rename() {
-  ConsoleLog('Build Step: Rename');
   for (const [from, to] of file_renamer) {
-    for (const path of new GlobScanner().scan(out_dir, `**/*${from}`).paths) {
-      await RenameFile(path, path.slice(0, path.lastIndexOf(from)) + to);
-      onLog(`Renamed: ${path}`);
+    for (const path_group of new GlobScanner().scan(out_dir, `**/*${from}`).path_groups) {
+      const path = path_group.path;
+      await RenameFile(path_group, new Path(path.slice(0, path.lastIndexOf(from)) + to));
+      onLog(`move: ${path}`);
     }
   }
 }
@@ -191,7 +181,7 @@ export async function buildStep_Rename() {
 export const on_log = new Broadcast<void>();
 export function onLog(data: string) {
   if (build_mode.silent === false) {
-    ConsoleLog(`[${new Date().toLocaleTimeString()}] > ${data}`);
+    ConsoleLogWithDate(`> ${data}`);
     on_log.send();
   }
 }